Fix table_combined; rename dissimilarity_heatmaps to distances
